backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.5
langchain-openai==0.0.6
tavily-python==0.3.0
pydantic==2.5.0
pydantic-settings==2.0.3
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
prometheus-client==0.19.0
pytest==7.4.3
httpx==0.25.2
